Safari-611.1.10/LayoutTests/js/dom/parse-error-external-script-in-eval.html

<html>
<head>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<script src="../resources/parse-error-external-script-in-eval.js"></script>
</head>
<body>
This tests that a line number and error message is printed for a parse error in a separate source file using eval().
</body>
</html>
