Safari-609.1.11.1/LayoutTests/js/dom/parse-error-external-script-in-eval-expected.txt

CONSOLE MESSAGE: line 2: SyntaxError: Unexpected identifier 'error'
This tests that a line number and error message is printed for a parse error in a separate source file using eval().
